Revert class naming to maintain compatibility
* Keep existing `*JacksonDeserializer` classes unchanged
* Add new `*Jackson3Deserializer` classes for Jackson 3 support
