+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/AdviceMessageJacksonDeserializer.java
@@ -16,20 +16,23 @@
 
 package org.springframework.integration.support.json;
 
-import tools.jackson.core.JacksonException;
-import tools.jackson.databind.DeserializationContext;
-import tools.jackson.databind.JsonNode;
+import java.io.IOException;
+
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonNode;
 
 import org.springframework.integration.message.AdviceMessage;
 import org.springframework.integration.support.MutableMessageHeaders;
 import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageHeaders;
 
 /**
  * The {@link MessageJacksonDeserializer} implementation for the {@link AdviceMessage}.
  *
- * @author Jooyoung Pyoung
+ * @author Artem Bilan
+ * @author Ngoc Nhan
  *
- * @since 7.0
+ * @since 4.3.10
  */
 public class AdviceMessageJacksonDeserializer extends MessageJacksonDeserializer<AdviceMessage<?>> {
 
@@ -42,9 +45,9 @@ public class AdviceMessageJacksonDeserializer extends MessageJacksonDeserializer
 
 	@Override
 	protected AdviceMessage<?> buildMessage(MutableMessageHeaders headers, Object payload, JsonNode root,
-			DeserializationContext ctxt) throws JacksonException {
-		Message<?> inputMessage = getMapper().readValue(root.get("inputMessage").traverse(ctxt), Message.class);
-		return new AdviceMessage<>(payload, headers, inputMessage);
+			DeserializationContext ctxt) throws IOException {
+		Message<?> inputMessage = getMapper().readValue(root.get("inputMessage").traverse(), Message.class);
+		return new AdviceMessage<>(payload, (MessageHeaders) headers, inputMessage);
 	}
 
 }
