+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/GenericMessageJacksonDeserializer.java
@@ -16,9 +16,10 @@
 
 package org.springframework.integration.support.json;
 
-import tools.jackson.core.JacksonException;
-import tools.jackson.databind.DeserializationContext;
-import tools.jackson.databind.JsonNode;
+import java.io.IOException;
+
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonNode;
 
 import org.springframework.integration.support.MutableMessageHeaders;
 import org.springframework.messaging.support.GenericMessage;
@@ -26,9 +27,9 @@ import org.springframework.messaging.support.GenericMessage;
 /**
  * The {@link MessageJacksonDeserializer} implementation for the {@link GenericMessage}.
  *
- * @author Jooyoung Pyoung
+ * @author Artem Bilan
  *
- * @since 7.0
+ * @since 4.3.10
  */
 public class GenericMessageJacksonDeserializer extends MessageJacksonDeserializer<GenericMessage<?>> {
 
@@ -41,7 +42,7 @@ public class GenericMessageJacksonDeserializer extends MessageJacksonDeserialize
 
 	@Override
 	protected GenericMessage<?> buildMessage(MutableMessageHeaders headers, Object payload, JsonNode root,
-			DeserializationContext ctxt) throws JacksonException {
+			DeserializationContext ctxt) throws IOException {
 		return new GenericMessage<Object>(payload, headers);
 	}
