+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/MessageJacksonDeserializer.java
@@ -16,18 +16,18 @@
 
 package org.springframework.integration.support.json;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-import tools.jackson.core.JacksonException;
-import tools.jackson.core.JsonParser;
-import tools.jackson.databind.DeserializationContext;
-import tools.jackson.databind.JavaType;
-import tools.jackson.databind.JsonNode;
-import tools.jackson.databind.ObjectMapper;
-import tools.jackson.databind.deser.std.StdNodeBasedDeserializer;
-import tools.jackson.databind.jsontype.TypeDeserializer;
-import tools.jackson.databind.type.TypeFactory;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JavaType;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.deser.std.StdNodeBasedDeserializer;
+import com.fasterxml.jackson.databind.jsontype.TypeDeserializer;
+import com.fasterxml.jackson.databind.type.TypeFactory;
 
 import org.springframework.integration.support.MutableMessageHeaders;
 import org.springframework.messaging.Message;
@@ -38,15 +38,15 @@ import org.springframework.util.Assert;
  *
  * @param <T> the message type.
  *
- * @author Jooyoung Pyoung
+ * @author Artem Bilan
  *
- * @since 7.0
+ * @since 4.3.10
  */
 public abstract class MessageJacksonDeserializer<T extends Message<?>> extends StdNodeBasedDeserializer<T> {
 
 	private static final long serialVersionUID = 1L;
 
-	private JavaType payloadType = TypeFactory.createDefaultInstance().constructType(Object.class);
+	private JavaType payloadType = TypeFactory.defaultInstance().constructType(Object.class);
 
 	private ObjectMapper mapper = new ObjectMapper();
 
@@ -70,20 +70,20 @@ public abstract class MessageJacksonDeserializer<T extends Message<?>> extends S
 
 	@Override
 	public Object deserializeWithType(JsonParser jp, DeserializationContext ctxt, TypeDeserializer td)
-			throws JacksonException {
+			throws IOException {
 
 		return super.deserialize(jp, ctxt);
 	}
 
 	@Override
-	public T convert(JsonNode root, DeserializationContext ctxt) throws JacksonException {
-		Map<String, Object> headers = this.mapper.readValue(root.get("headers").traverse(ctxt),
-				TypeFactory.createDefaultInstance().constructMapType(HashMap.class, String.class, Object.class));
-		Object payload = this.mapper.readValue(root.get("payload").traverse(ctxt), this.payloadType);
+	public T convert(JsonNode root, DeserializationContext ctxt) throws IOException {
+		Map<String, Object> headers = this.mapper.readValue(root.get("headers").traverse(),
+				TypeFactory.defaultInstance().constructMapType(HashMap.class, String.class, Object.class));
+		Object payload = this.mapper.readValue(root.get("payload").traverse(), this.payloadType);
 		return buildMessage(new MutableMessageHeaders(headers), payload, root, ctxt);
 	}
 
 	protected abstract T buildMessage(MutableMessageHeaders headers, Object payload, JsonNode root,
-			DeserializationContext ctxt) throws JacksonException;
+			DeserializationContext ctxt) throws IOException;
 
 }
